[CSS] Generate contiguous responsive CSS vars
Ensures the CSS vars set via `style` props will override the
responsive-props SASS mixin which forces them to have a value of
`initial`.

Forcing to `initial` is required when components are nested to avoid an
outer component accidentally setting a value for a nested component by
specifying a CSS variable (CSS vars are globally scoped).

Adding this `forceContiguous` flag means we can be sure we're setting
exactly what the user asked for when passing in their prop, and not
`initial` or a parent component's value.